test: Seclude core resources setup
