+++ node/packages/aws-lambda-otel-extension/test/integration/index.test.js
@@ -3,22 +3,18 @@
 const { expect } = require('chai');
 
 const path = require('path');
-const fsp = require('fs').promises;
 const wait = require('timers-ext/promise/sleep');
 const { CloudWatchLogs } = require('@aws-sdk/client-cloudwatch-logs');
 const { Lambda } = require('@aws-sdk/client-lambda');
-const { IAM } = require('@aws-sdk/client-iam');
-const { STS } = require('@aws-sdk/client-sts');
 const log = require('log').get('test');
-const buildLayer = require('../../scripts/lib/build');
 const resolveDirZipBuffer = require('../utils/resolve-dir-zip-buffer');
 const normalizeOtelAttributes = require('../utils/normalize-otel-attributes');
 const ensureNpmDependencies = require('../../scripts/lib/ensure-npm-dependencies');
 const awsRequest = require('./aws-request');
 const basename = require('./basename');
 const cleanup = require('./cleanup');
+const createCoreResources = require('./create-core-resources');
 
-const layerFilename = path.resolve(__dirname, '../../dist/extension.zip');
 const fixturesDirname = path.resolve(__dirname, '../fixtures/lambdas');
 
 describe('integration', function () {
@@ -196,87 +192,6 @@ describe('integration', function () {
 
   before(async () => {
     ensureNpmDependencies('test/fixtures/lambdas');
-    log.notice('Creating %s', basename);
-
-    config.accountId = (await awsRequest(STS, 'getCallerIdentity')).Account;
-    const createLayer = async () => {
-      if (!process.env.TEST_LAYER_FILENAME) {
-        log.info('Building layer');
-        await buildLayer(layerFilename);
-      }
-
-      log.info('Publishing layer (%s) to AWS', process.env.TEST_LAYER_FILENAME || layerFilename);
-      await awsRequest(Lambda, 'publishLayerVersion', {
-        LayerName: basename,
-        Content: { ZipFile: await fsp.readFile(process.env.TEST_LAYER_FILENAME || layerFilename) },
-      });
-      log.info('Resolving layer ARN');
-      config.layerArn = (
-        await awsRequest(Lambda, 'listLayerVersions', { LayerName: basename })
-      ).LayerVersions.shift().LayerVersionArn;
-      log.info('Layer ready %s', config.layerArn);
-    };
-    const createRole = async () => {
-      log.info('Creating IAM role and policy');
-      const [
-        {
-          Role: { Arn: roleArn },
-        },
-        {
-          Policy: { Arn: policyArn },
-        },
-      ] = await Promise.all([
-        awsRequest(IAM, 'createRole', {
-          RoleName: basename,
-          AssumeRolePolicyDocument: JSON.stringify({
-            Version: '2012-10-17',
-            Statement: [
-              {
-                Effect: 'Allow',
-                Principal: { Service: ['lambda.amazonaws.com'] },
-                Action: ['sts:AssumeRole'],
-              },
-            ],
-          }),
-        }).catch((error) => {
-          if (error.Code === 'EntityAlreadyExists') {
-            log.notice('IAM role already exists');
-            return { Role: { Arn: `arn:aws:iam::${config.accountId}:role/${basename}` } };
-          }
-          throw error;
-        }),
-        awsRequest(IAM, 'createPolicy', {
-          PolicyName: basename,
-          PolicyDocument: JSON.stringify({
-            Version: '2012-10-17',
-            Statement: [
-              {
-                Effect: 'Allow',
-                Action: ['logs:CreateLogStream', 'logs:CreateLogGroup'],
-                Resource: `arn:*:logs:*:*:log-group:/aws/lambda/${basename}*:*`,
-              },
-              {
-                Effect: 'Allow',
-                Action: ['logs:PutLogEvents'],
-                Resource: `arn:*:logs:*:*:log-group:/aws/lambda/${basename}*:*:*`,
-              },
-            ],
-          }),
-        }).catch((error) => {
-          if (error.Code === 'EntityAlreadyExists') {
-            log.notice('IAM policy already exists');
-            return { Policy: { Arn: `arn:aws:iam::${config.accountId}:policy/${basename}` } };
-          }
-          throw error;
-        }),
-      ]);
-      log.info('Attaching IAM policy to role');
-      await awsRequest(IAM, 'attachRolePolicy', { RoleName: basename, PolicyArn: policyArn });
-      log.info('Attached IAM policy to role');
-      config.roleArn = roleArn;
-      config.policyArn = policyArn;
-    };
-
     const createFunctions = async () => {
       await Promise.all(
         Array.from(functionsConfig, async function self([handlerModuleName, { creationOptions }]) {
@@ -333,8 +248,7 @@ describe('integration', function () {
 
     [lambdasCodeZipBuffer] = await Promise.all([
       resolveDirZipBuffer(fixturesDirname),
-      createLayer(),
-      createRole(),
+      createCoreResources(config),
     ]);
     await createFunctions();
   });
